# openapi.yaml
openapi: 3.0.3
info:
  title: Canvas Connect API
  description: |
    Backend contract expected by the `frontend/` client (see `frontend/src/lib/api.ts`,
    `frontend/src/lib/types.ts` and `frontend/src/lib/realtime.ts`). This file is the
    source of truth for the REST surface the backend must implement; the frontend
    currently falls back to an in-memory mock (`frontend/src/lib/mock-backend.ts`)
    whenever `VITE_API_BASE_URL` is not set.

    ## Authentication

    Authenticated requests carry `Authorization: Bearer <token>`. Two kinds of
    bearer token exist:

    - **User tokens** — interviewers. `POST /v1/auth/magic-link` is documented
      here as the primary flow, but this backend has no email provider, so it
      always reports `{ "sent": true }` without issuing a token. The practical
      way to obtain a user token is `POST /v1/auth/login` (email + password;
      passwords are stored hashed, never in plaintext).
    - **Participant tokens** — guests. Returned as `collaboration_token` by
      `POST /v1/join/{token}` and reissued by `GET /v1/sessions/{id}/canvas`.
      Usable both as the WebSocket join credential and as a normal bearer
      token on the REST endpoints the participant's role is allowed to call.

    Note: the reference frontend client (`frontend/src/lib/api.ts`) currently
    calls `fetch(..., { credentials: "include" })` and sends no `Authorization`
    header — it was written against an earlier cookie-based draft of this
...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
  /v1/auth/magic-link:
    post:
      tags: [auth]
      summary: Request a magic-link sign-in email
      operationId: requestMagicLink
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicLinkRequest"
      responses:
        "200":
          description: Email dispatched (always returns sent:true to avoid leaking account existence)
          content:
            application/json:
              schema:
                type: object
                required: [sent]
                properties:
                  sent:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/auth/login:
    post:
      tags: [auth]
      summary: Exchange email + password for a bearer token
      operationId: login
      description: |
        Not part of the original illustrative API surface in docs/specs.md —
        added because this backend has no email provider to complete the
        magic-link flow above. Passwords are verified against a bcrypt hash;
        the plaintext password is never stored.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Email or password is incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/me:
    get:
      tags: [auth]
      summary: Get the currently authenticated user
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/sessions/{id}/guest-links:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      tags: [guest-links]
      summary: Create or rotate a guest link
      operationId: createGuestLink
      description: |
        Owner, or interviewer if granted control. Creating a new link revokes any
        previously active link for the session (rotation semantics).
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuestLinkRequest"
      responses:
        "201":
          description: Guest link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestLink"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [guest-links]
      summary: List active guest links for a session
      operationId: listGuestLinks
      description: Owner or interviewer. Revoked links are excluded.
      responses:
        "200":
          description: Active guest links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GuestLink"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/sessions/{id}/guest-links/{linkId}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
      - name: linkId
        in: path
        required: true
        schema:
          type: string
    delete:
      tags: [guest-links]
      summary: Revoke a guest link
      operationId: revokeGuestLink
      description: Owner only.
      responses:
        "204":
          description: Link revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/join/{token}:
    parameters:
      - name: token
        in: path
        required: true
        description: Bearer token embedded in the guest link URL.
        schema:
          type: string
    get:
      tags: [join]
      summary: Lobby preflight — validate a guest link without joining
      operationId: previewJoin
      description: Public. Used to render the join page before the candidate submits a display name.
      security: []
      responses:
        "200":
          description: |
            Always 200; `joinable` is false with a human-readable `reason` for
            revoked, expired, archived, or capacity-limited links, or an invalid token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinPreview"
    post:
      tags: [join]
      summary: Validate token and create a guest participation
      operationId: joinSession
      description: |
        Public (pre-auth). On success the backend must set the session cookie for
        the newly created guest participant, in addition to returning a short-lived
        `collaboration_token` used for the WebSocket handshake.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRequest"
      responses:
        "201":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Link revoked, expired, archived, or at max capacity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/sessions/{id}/canvas:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [canvas]
      summary: Obtain the canvas snapshot and collaboration credentials
      operationId: getCanvas
      description: Any authenticated participant of the session (owner, interviewer, candidate, or observer).
      responses:
        "200":
          description: Canvas snapshot plus WebSocket handoff
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvasGetResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [canvas]
      summary: Autosave fallback when the collaboration socket is down
      operationId: saveCanvas
      description: |
        Any participant with edit permission (owner/interviewer always; candidate
        only while `candidate_editing_enabled` is true). The primary write path is
        the WebSocket `document_update` message — this endpoint exists purely as a
        fallback so state is not lost while disconnected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CanvasDocument"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                type: object
                required: [saved_at]
                properties:
                  saved_at:
                    type: string
                    format: date-time
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        User token from POST /v1/auth/login, or participant (collaboration)
        token from POST /v1/join/{token} or GET /v1/sessions/{id}/canvas.
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but not permitted to perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource does not exist (or is not visible to the caller)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: not_found
        message:
          type: string
          example: Session not found
    Role:
      type: string
      enum: [owner, interviewer, candidate, observer]
    SessionState:
      type: string
      enum: [draft, live, ended, archived]
    ConnectionState:
      type: string
      enum: [connected, reconnecting, offline]
    User:
      type: object
      required: [id, email, display_name, organization_id, created_at]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        organization_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    Participant:
      type: object
      required: [id, session_id, user_id, display_name, role, color, joined_at, left_at]
      properties:
        id:
          type: string
        session_id:
          type: string
        user_id:
          type: string
          nullable: true
          description: Null for guest (candidate) participants.
        display_name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        color:
          type: string
          description: Presence color assigned from the server palette.
        joined_at:
          type: string
          format: date-time
        left_at:
          type: string
          format: date-time
          nullable: true
        connection_state:
          $ref: "#/components/schemas/ConnectionState"
        cursor:
          type: object
          nullable: true
          required: [x, y]
          properties:
            x:
              type: number
            y:
              type: number
        selection:
          type: array
          items:
            type: string
    InterviewSession:
      type: object
      required:
        - id
        - owner_user_id
        - title
        - prompt
        - state
        - candidate_editing_enabled
        - cursors_visible
        - scheduled_at
        - started_at
        - ended_at
        - created_at
        - updated_at
        - participants
      properties:
        id:
          type: string
        owner_user_id:
          type: string
        title:
          type: string
        prompt:
          type: string
        state:
          $ref: "#/components/schemas/SessionState"
        candidate_editing_enabled:
          type: boolean
        cursors_visible:
          type: boolean
        scheduled_at:
          type: string
          format: date-time
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        ended_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Participant"
    GuestLink:
      type: object
      required: [id, session_id, url, role_granted, expires_at, max_uses, revoked_at, created_at]
      properties:
        id:
          type: string
        session_id:
          type: string
        url:
          type: string
          format: uri
        role_granted:
          $ref: "#/components/schemas/Role"
        expires_at:
          type: string
          format: date-time
          nullable: true
        max_uses:
          type: integer
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    MagicLinkRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    TokenResponse:
      type: object
      required: [access_token, token_type]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
...
    CreateGuestLinkRequest:
      type: object
      properties:
        role_granted:
          $ref: "#/components/schemas/Role"
        expires_at:
          type: string
          format: date-time
          nullable: true
        max_uses:
          type: integer
          nullable: true
    JoinPreview:
      type: object
      required: [session_title, joinable]
      properties:
        session_title:
          type: string
        joinable:
          type: boolean
        reason:
          type: string
          description: Present when joinable is false.
    JoinRequest:
      type: object
      required: [display_name]
      properties:
        display_name:
          type: string
    JoinResponse:
      type: object
      required: [session, participant, collaboration_token]
      properties:
        session:
          $ref: "#/components/schemas/InterviewSession"
        participant:
          $ref: "#/components/schemas/Participant"
        collaboration_token:
          type: string
          description: Short-lived credential used for the WebSocket join_room handshake.
    CanvasGetResponse:
      type: object
      required: [document, collaboration_token, websocket_url]
      properties:
        document:
          $ref: "#/components/schemas/CanvasDocument"
        collaboration_token:
          type: string
        websocket_url:
          type: string
          format: uri
    ComponentType:
      type: string
      enum:
        - service
        - rounded
        - boundary
        - generic
        - sql-db
        - nosql-db
        - cache
        - object-storage
        - warehouse
        - queue
        - stream
        - pubsub
        - client
        - mobile-client
        - api-gateway
        - load-balancer
        - cdn
        - external-api
        - server
        - worker
        - function
        - container
        - llm
        - embedding
        - vector-db
        - agent
    BaseElement:
      type: object
      required: [id, kind, x, y, z, parent_id, created_by, created_at, updated_at]
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [component, shape, text, sticky, stroke, connector]
        x:
          type: number
        y:
          type: number
        z:
          type: number
        parent_id:
          type: string
          nullable: true
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BoxElement:
      description: kind is one of component, shape, text, sticky.
      allOf:
        - $ref: "#/components/schemas/BaseElement"
        - type: object
          required: [width, height, label]
          properties:
            width:
              type: number
            height:
              type: number
            label:
              type: string
            description:
              type: string
            componentType:
              $ref: "#/components/schemas/ComponentType"
            shape:
              type: string
              enum: [rect, ellipse]
            color:
              type: string
    StrokeElement:
      allOf:
        - $ref: "#/components/schemas/BaseElement"
        - type: object
          required: [points, color, width, tool]
          properties:
            points:
              type: array
              description: Flat [x, y, x, y, ...] pairs in canvas space.
              items:
                type: number
            color:
              type: string
            width:
              type: number
            tool:
              type: string
              enum: [pen, highlighter]
    ConnectorEndpoint:
      type: object
      required: [elementId, x, y]
      properties:
        elementId:
          type: string
          nullable: true
        x:
          type: number
        y:
          type: number
    ConnectorElement:
      allOf:
        - $ref: "#/components/schemas/BaseElement"
        - type: object
          required: [from, to, style, dashed, color, strokeWidth, arrowStart, arrowEnd]
          properties:
            from:
              $ref: "#/components/schemas/ConnectorEndpoint"
            to:
              $ref: "#/components/schemas/ConnectorEndpoint"
            style:
              type: string
              enum: [straight, elbow, curved]
            dashed:
              type: boolean
            color:
              type: string
            strokeWidth:
              type: number
            arrowStart:
              type: boolean
            arrowEnd:
              type: boolean
            label:
              type: string
    CanvasElement:
      oneOf:
        - $ref: "#/components/schemas/BoxElement"
        - $ref: "#/components/schemas/StrokeElement"
        - $ref: "#/components/schemas/ConnectorElement"
      discriminator:
        propertyName: kind
        mapping:
          component: "#/components/schemas/BoxElement"
          shape: "#/components/schemas/BoxElement"
          text: "#/components/schemas/BoxElement"
          sticky: "#/components/schemas/BoxElement"
          stroke: "#/components/schemas/StrokeElement"
          connector: "#/components/schemas/ConnectorElement"
    CanvasDocument:
      type: object
      required: [id, session_id, schema_version, latest_operation_cursor, updated_at, elements]
      properties:
        id:
          type: string
        session_id:
          type: string
        schema_version:
          type: integer
        latest_operation_cursor:
          type: integer
        updated_at:
          type: string
          format: date-time
        elements:
          type: array
          items:
            $ref: "#/components/schemas/CanvasElement"

    # --- Realtime protocol (documentation only; not bound to an HTTP path) ---
...